requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
python-dotenv
pillow
supabase
